refactor(subjects): nest remaining root types into folders by sub-namespace
Finishes the namespace reorg series (after #899-#903) for RedditPodcastPoster.Subjects.

- Providers/: CachedSubjectProvider, ICachedSubjectProvider, RecycledFlareIdProvider, IRecycledFlareIdProvider

- Repositories/: SubjectRepository

- Enrichers/: SubjectEnricher, ISubjectEnricher

- Matching/: SubjectMatcher, ISubjectMatcher

- Services/: SubjectService, ISubjectService

- Cleansing/: SubjectCleanser, ISubjectCleanser

- Categorisation/: Categoriser, ICategoriser, RecentPodcastEpisodeCategoriser, IRecentPodcastEpisodeCategoriser, CategorisePodcastLogger, CategoriseEpisodeDelta

Root goes from 19 .cs files (59% of 32) to 0. Existing Extensions/Factories/HashTags/Models folders left untouched. Fixed all consumer usings across Console-Apps, Cloud/Api, Cloud/Indexer, FunctionHost.Tests, Discovery, Indexing, UrlSubmission, PodcastServices.Tests, and Subjects.Tests. Public type names and DI extension method names unchanged.
