+++ Class-Libraries/RedditPodcastPoster.Subjects.Tests/SubjectServiceTests.cs
@@ -3,6 +3,7 @@ using FluentAssertions;
 using Moq.AutoMock;
 using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Persistence.Abstractions.Providers;
+using RedditPodcastPoster.Subjects.Services;
 
 namespace RedditPodcastPoster.Subjects.Tests;
