+++ Class-Libraries/RedditPodcastPoster.Subjects/Categorisation/Categoriser.cs
@@ -1,8 +1,9 @@
 ﻿using Microsoft.Extensions.Logging;
 using RedditPodcastPoster.Models;
+using RedditPodcastPoster.Subjects.Enrichers;
 using RedditPodcastPoster.Subjects.Models;
 
-namespace RedditPodcastPoster.Subjects;
+namespace RedditPodcastPoster.Subjects.Categorisation;
 
 public class Categoriser(
     ISubjectEnricher subjectEnricher,
