+++ Class-Libraries/RedditPodcastPoster.Subjects/Categorisation/ICategoriser.cs
@@ -1,6 +1,6 @@
 ﻿using RedditPodcastPoster.Models;
 
-namespace RedditPodcastPoster.Subjects;
+namespace RedditPodcastPoster.Subjects.Categorisation;
 
 public interface ICategoriser
 {
