+++ Class-Libraries/RedditPodcastPoster.Subjects/Enrichers/ISubjectEnricher.cs
@@ -1,7 +1,7 @@
 ﻿using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Subjects.Models;
 
-namespace RedditPodcastPoster.Subjects;
+namespace RedditPodcastPoster.Subjects.Enrichers;
 
 public interface ISubjectEnricher
 {
