+++ Class-Libraries/RedditPodcastPoster.Subjects/Matching/ISubjectMatcher.cs
@@ -1,7 +1,7 @@
 ﻿using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Subjects.Models;
 
-namespace RedditPodcastPoster.Subjects;
+namespace RedditPodcastPoster.Subjects.Matching;
 
 public interface ISubjectMatcher
 {
