+++ Class-Libraries/RedditPodcastPoster.Subjects/Providers/ICachedSubjectProvider.cs
@@ -1,5 +1,5 @@
 ﻿using RedditPodcastPoster.Persistence.Abstractions.Providers;
 
-namespace RedditPodcastPoster.Subjects;
+namespace RedditPodcastPoster.Subjects.Providers;
 
 public interface ICachedSubjectProvider : ISubjectsProvider;
