+++ Class-Libraries/RedditPodcastPoster.Subjects/Repositories/SubjectRepository.cs
@@ -6,7 +6,7 @@ using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Persistence.Abstractions.Providers;
 using RedditPodcastPoster.Persistence.Abstractions.Repositories;
 
-namespace RedditPodcastPoster.Subjects;
+namespace RedditPodcastPoster.Subjects.Repositories;
 
 public class SubjectRepository(
     Container subjectsContainer,
