+++ Class-Libraries/RedditPodcastPoster.Subjects/Services/ISubjectService.cs
@@ -1,7 +1,7 @@
 ﻿using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Subjects.Models;
 
-namespace RedditPodcastPoster.Subjects;
+namespace RedditPodcastPoster.Subjects.Services;
 
 public interface ISubjectService
 {
