+++ Cloud/Indexer/Categoriser.cs
@@ -3,7 +3,7 @@ using Microsoft.DurableTask;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using RedditPodcastPoster.Configuration;
-using RedditPodcastPoster.Subjects;
+using RedditPodcastPoster.Subjects.Categorisation;
 
 namespace Indexer;
